docker/requirements.txt
Flask == 0.12.1
SQLAlchemy == 1.2.*
Flask-SQLAlchemy == 2.3.*
python-dateutil  == 2.7.*
phpserialize == 1.3.*
Flask-Injector == 0.10.*
injector == 0.13.*
mimerender == 0.6.*
sqlalchemy-utils == 0.33.*
psycopg2-binary == 2.7.*